Filter more devices name and use kube-node group
The device name is now filtered removing the prefix '/dev/'.
If the string 'mapper' is still present in the device, we currently
raise an AssertionError, waiting for more input on how ansible see the
device-mapper managed devices.

Also, the groups 'kube-node' is used for the storage check instead of
iterating on all the hosts.

Issue #231